requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
psycopg2-binary==2.9.7